Rename shared => shared_conversion_functions
