+++ src/conversion/area/acres.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const area = require('./_area');
 
 module.exports = {
