+++ src/conversion/distance/nautical_mile.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const distance = require('./_distance');
 
 module.exports = {
