+++ src/conversion/force/pound_force.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const force = require('./_force');
 
 module.exports = {
