+++ src/conversion/fuel_economy/_fuel_economy.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 
 function fuelEconomyMap(imperialInputs, metricTransform) {
   const kmL = imperialInputs.map(metricTransform);
