+++ src/conversion/pressure/psi.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const pressure = require('./_pressure');
 
 module.exports = {
