+++ src/conversion/shared_conversion_functions.js
