+++ src/conversion/temperature/fahrenheit.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const temperature = require('./_temperature');
 
 module.exports = {
