+++ src/conversion/torque/lb_per_in.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const torque = require('./_torque');
 
 module.exports = {
