+++ src/conversion/velocity/mph.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const velocity = require('./_velocity');
 
 module.exports = {
