+++ src/conversion/volume/gallons.js
@@ -1,4 +1,4 @@
-const shared = require('../shared');
+const shared = require('../shared_conversion_functions');
 const volume = require('./_volume');
 
 module.exports = {
